refactor(p2): extract ICloud into sibling page (stage 3 of 4)
SettingExperimentalICloudPage + ICloudLoginDialog (private) +
ICLOUD_WEBVIEW_USER_AGENT private const moved from
SettingExperimentalPage.kt into new sibling SettingExperimentalICloudPage.kt.

Two non-contiguous source ranges in the post-stage2 coordinator
(parent 875cccb8): L153-310 (ICloudPage), L1247-1313 (ICloudLoginDialog +
const + intervening blank). Combined in the new file in source order with
their natural blank separator preserved.

Byte-equal: bodies and signatures unchanged.
Visibility unchanged.

Coordinator strip: 1398 -> 1168 (-230 with import cleanup).
Imports trimmed in coordinator: 5 orphans
(ICLOUD_CHINA_LOGIN_URL, ICLOUD_GLOBAL_LOGIN_URL, ICloudDriveManager,
WebView, rememberWebViewState).

Pre-existing dead `base64Encode` import left untouched (was never used in
this file even before the refactor started; cdbdf159 baseline confirms).

Baseline 6/463 unit tests preserved.
